Complete code for vistualization
This contains all the code I have written to create this visualization web based project for the data. It was a good practice for the data science. The core library used in this project was Steamlit.
If you want to run this project, pip install the streamlit library along with others mentioned in the code in the  beginnings like Pandas and Numpy, then run the code form the terminal

steamlit run app.py
Make sure to keep both code and the csv file in the same folder.